Create downloading_data
This code block utilizes the Landsatxplore API to access data from the NASA Landsat Earth Explorer website. It prompts the user to enter their Earth Explorer credentials, and based on the specified date, location, and cloud cover criteria defined in the program, it generates a list of all available data files. The user is then prompted to download one or more of these files as desired.