Create raster_masker
This module, raster_masker, was created to streamline the process of masking raster files using shapefiles. It defines a RasterMasker class with a mask_raster method that masks input raster files based on shapefile geometries and saves the results. This functionality is essential for isolating specific areas of interest within large raster datasets.